requirements.txt
flask
gunicorn
python-dotenv
flask-mail
requests